pyproject.toml
[project]
name = "agenticlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.4.8",
    "python-dotenv>=1.1.1",
]
